Nmap now runs with the `http-methods` script that will provide the list of supported HTTP methods.
In some cases (e.g. confluence), `http-methods` does not return any results (the server does not handle or answer to `OPTIONS` request). In that case, we use the results of `fingerprint-strings` to try to best-effort match missed web services.

This change can be expensive as it means that more services are going to be flagged as web service and will overall increase the
running time of Tsunami. But it will also increase coverage greatly.

PiperOrigin-RevId: 592197503
Change-Id: If1e6332700316d073846bc19eef7ca8e9dca2d15